requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3